Changes to handle ST foundry
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@2570 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/tool/user/dialogs/CellLists.java
@@ -214,7 +214,7 @@ public class CellLists extends EDialog
 		if (maxlen < 0) line += "\t"; else line += " ";
 
 		boolean goodDRC = false;
-        byte activeBits = DRC.getActiveBits();
+        int activeBits = DRC.getActiveBits();
 		Date lastGoodDate = DRC.getLastDRCDateBasedOnBits(cell, activeBits);
 		if (!Main.getDebug() && lastGoodDate != null && cell.getRevisionDate().before(lastGoodDate)) goodDRC = true;
 		if (goodDRC) line += "D"; else line += " ";
